+++ Runtime/Simulation/Environment/SimulatedTrackedImage.cs
@@ -89,7 +89,9 @@ namespace UnityEngine.XR.Simulation
 
         TrackableId GenerateTrackableID()
         {
-#if UNITY_6000_4_OR_NEWER
+#if UNITY_6000_5_OR_NEWER
+            var unsignedInstanceId = gameObject.GetEntityId().GetRawData();
+#elif UNITY_6000_4_OR_NEWER
             var unsignedInstanceId = (ulong)Convert.ToInt64(gameObject.GetEntityId());
 #else
             var unsignedInstanceId = (ulong)Convert.ToInt64(gameObject.GetInstanceID());
@@ -99,7 +101,9 @@ namespace UnityEngine.XR.Simulation
 
         Guid GenerateSourceImageId()
         {
-#if UNITY_6000_4_OR_NEWER
+#if UNITY_6000_5_OR_NEWER
+            var unsignedInstanceId = gameObject.GetEntityId().GetRawData();
+#elif UNITY_6000_4_OR_NEWER
             var unsignedInstanceId = (ulong)Convert.ToInt64(gameObject.GetEntityId());
 #else
             var unsignedInstanceId = (ulong)Convert.ToInt64(gameObject.GetInstanceID());